refactor: move test packages and data to internal
