+++ internal/testhelper/exe/exe_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/Nivl/git-go/testhelper/exe"
+	"github.com/Nivl/git-go/internal/testhelper/exe"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
