+++ plumbing/packfile/packindex_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 
 	"github.com/Nivl/git-go"
+	"github.com/Nivl/git-go/internal/testhelper"
 	"github.com/Nivl/git-go/plumbing"
 	"github.com/Nivl/git-go/plumbing/packfile"
-	"github.com/Nivl/git-go/testhelper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/xerrors"
